Adding a receive request type

+++ types/message.go
@@ -4,7 +4,7 @@ import "gorm.io/gorm"
 
 type Message struct {
 	gorm.Model
-    QueueName string `gorm:"index"`
-    Partition int16 `gorm:"index"`
+	QueueName string `gorm:"index"`
+	Partition int    `gorm:"index"`
 	Data      []byte
 }
